+++ src/cpp/linear_elasticity.h
@@ -26,6 +26,12 @@ namespace linearElasticity{
 
     errorOut formReferenceStiffnessTensor( const floatVector &parameters, floatMatrix &stiffnessTensor );
 
+    errorOut formReferenceStiffnessTensor( const floatMatrix &directionCosines, const floatVector &parameters,
+                                           floatMatrix &stiffnessTensor );
+
+    errorOut formReferenceStiffnessTensor( const floatVector &bungeEulerAngles, const floatVector &parameters,
+                                           floatMatrix &stiffnessTensor );
+
     errorOut evaluateEnergy( const floatVector &chi, const floatVector &parameters, floatType &energy );
 
     errorOut evaluateEnergy( const floatVector &chi, const floatVector &parameters, floatType &energy, floatVector &cauchyStress );